Crate smashline

Re-exports

pub use smash as skyline_smash;

pub use smash_rs;

pub use locks;

Modules

api

Structs

Agent

AgentEntry

CalcParam

CheckAttack

CheckDamage

Costume

End

Exec

ExecStop

Exit

FixCamera

FixPosSlow

Hash40

The central type of the crate, representing a string hashed using the hash40 algorithm Hash40 is a combination of a crc32 checksum and string length appended to the top bits

Init

L2CAgentBase

L2CFighterBase

L2CFighterCommon

L2CValue

L2CWeaponCommon

LeaveStop

Main

MapCorrection

NotifyEventGimmick

OnChangeLr

Post

Pre

StringFFI

Variadic

Enums

Acmd

BattleObjectCategory

LuaConst

ObjectEvent

Priority

StatusLine

StringFFIError

Traits

AsHash40

IntoLuaConst

StatusLineMarker

Functions

acmd_stub⚠

add_param_object

clone_weapon

original_acmd

original_status

update_weapon_count

whitelist_kirby_copy_article

Type Aliases

AcmdFunction

StateFunction

Attribute Macros

acmd

event

in_target

line

new_status

status

unwindable